Add support for bind mounting directories inside the chroot.
This is useful for cases where large volumes of data outside the workspace might be wanted inside it - for example, downloading a full set of Android SDK tarballs takes about 16GB, and it's much easier to have a shared directory available for that cache to be shared between jobs.

This isn't supported by Mock - I'd welcome a smarter way to mask the option on Mock builds, but I couldn't figure out the correct Jelly conditional to only show the textbox for Pbuilder-type chroots.